final task
The folder contains two python files which has to be executed at the same time to accomplish the task. Save the files on the PC in same folders. From Terminal or Shell, execute the lidar_control file first then execute the lane_control file. The lidar_status.txt file holds the status value of lidar. The camera will follow the white line on the track and the car stop if an obstracle is in front of the car in the range of 18 cm.